OMAP3: HS: No Recovery From Warm Reset While In Off Mode
i537: A deadlock situation occurs if a warm reset occurs during Off
mode or after wakeup from Off but before the ROM code execution is
complete on an HS device.

Workaround: If sys_nreswarm is used as an OMAP input and is potentially
asserted while OMAP is in Off mode then the only way to recover is to
generate a power on reset. This can be programmed in the warm_reset
sequence if the TWL4030/5030 PMIC is used.

This patch modifies the warm reset sequence accordingly.  It also
corrects the delay in the All resource step to match that indicated in
the document "Configuring TWL5030 for Warm Reset Sequences" (SWCA082).

Change-Id: Idb26c44ec9e72411244966d8d2fd57a535641355
Signed-off-by: Steve Schefter <sschefter@ti.com>
Signed-off-by: Deepak Muddegowda <x0171695@ti.com>