encore: configure Vmmc2 as fixed voltage regulator instead of via PMIC
It appears that encore's PMIC doesn't have a Vmmc2 regulator, but we've
been attempting to set one up in the TWL anyway -- presumably for the
side effect of telling the MMC subsystem the input voltage it has
available.  The "correct" way of doing this is to set up a "fixed
regulator" which tells the kernel the voltage of the Vmmc2 supply.